requirements.txt
fastapi
uvicorn
gunicorn  # <-- Important for production
streamlit
psycopg2-binary
sentence-transformers
faiss-cpu
groq
pypdf
beautifulsoup4
requests
python-dotenv
langchain
langchain-groq
langchain-community  # For FAISS and other integrations
langchain-huggingface